// utils/types.ts
export interface iEnvVariables {
    AUTO_PLAYWRIGHT_DEBUG: boolean;
    BASE_URL: string;
    BASE_FIRM_NAME: string;
    BASE_MEASUREMENT_DATE: string;
    HEADLESS_DEBUG: boolean;
    OPENAI_API_KEY: string;
    SCALAR_DASH409A_PASSWORD: string;
    SCALAR_DASH409A_USERNAME: string;
    SCALAR_PASSWORD: string;
    SCALAR_USERNAME: string;
    SLACK_OAUTH_TOKEN: string;
    QATOUCH_DOMAIN: string;
    QATOUCH_API_TOKEN: string;
    QATOUCH_PROJECT_KEY: string;
    QATOUCH_TESTRUN_ID: string;
};